requirements.txt
# sctp-probe Python dependencies
# Pin exact versions to ensure reproducible installs in CI and WSL2.
# Update with: pip install -r requirements.txt && pip freeze > requirements.lock

# Web framework
fastapi==0.115.6
uvicorn[standard]==0.32.1

# Data validation (pulled in by fastapi, pinned explicitly)
pydantic==2.10.3

# SBc-AP ASN.1 APER decode/encode
# NOTE: pycrate 0.4.10 does not exist on PyPI. Verified available version: 0.7.11.
# The SBc-AP (TS 29.168) module is NOT shipped with pycrate — it must be compiled
# from ASN.1 source files (see specs/sbcap/ and specs/SbcAP_gen.py).
pycrate==0.7.11

# SCTP Linux kernel socket wrapper (Linux only)
# Install manually on non-Linux if needed for import compatibility:
#   pip install pysctp  (will fail to compile on Windows — that is expected)
# NOTE: pysctp 0.6.2 does not exist on PyPI. Verified available version: 0.7.3.
pysctp==0.7.3

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.27.2       # required by FastAPI TestClient
